Fix for infinite loop, etc. (#3)
* Fixed a couple of small bugs surrounding backslash handling in a
format string.  First, if someone wanted a literal backslash in there,
with no '{', '}' or '\' after, the algorithm would hang, because the
index wasn't advancing when a backslash wasn't followed by one of
those things (backslash forever!).  Second, if that backslash were the
last character of the input, then you'd get a nice array index out of
bounds error.

* Update src/formatstr.nim

* Update src/formatstr.nim